backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
celery[redis]>=5.4.0
redis>=5.0.0
python-multipart>=0.0.9
httpx>=0.27.0
sqlmodel>=0.0.22
google-genai>=1.0.0
Pillow>=10.0.0
